Attach Cache to individual pages. Implement relative caching
